kmscube: add write to png option to offscreen
Add an option to write frames to png files. libpng was already
an optional dependency to kmscube and texturator already had that
feature, so we can reuse most of that.
This is mainly useful in the offscreen mode to view the rendered
frames when using a machine without an accessible display, so for
now limit this feature to offscreen.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>