kmscube: Cast normal matrix from modelview matrix in shader
The operation of taking the 3x3 matrix starting at 0,0 from a 4x4 matrix
is common enough that it is the default action when casting mat4 to mat3.
So instead of manually building out a 3x3 normal matrix from the modelview
and passing that into the shader, simply have the shader cast the normal
matrix from the already available modelview matrix.

Signed-off-by: Andrew Davis <afd@ti.com>