kmscube: Get attribute locations dynamically over static bindings
While both specifying the binding location and letting the compiler pick
the location do work, setting the binding manually often leads to bugs
like found in previous patches. Since we do not need to force aliasing
here we can switch to letting the compiler pick. This also allows for
combining compilation and linking as no in-between step of specifying
the binding location is needed.

Signed-off-by: Andrew Davis <afd@ti.com>