kmscube: Move cube vertex data to common location
All the cubes can share a common set of vertex data which simplifies
each cube and saves program data. Factor out the common cube vertex
data into a common location.

Also interleave the vertex data which allows for loading this data
to the GPU with a single glBufferData() call. This should also make
the vertex fetching slightly more efficient by keeping the data for
each vertex closer in memory (not that this really matters much, but
it is nice to have example code like this do things the right way).

Lastly, label the data components to more easily see what is going
into each vertex while defining the data structure.

Signed-off-by: Andrew Davis <afd@ti.com>