cube-gears: Do not use renderbuffer texture targets
Using a normal texture as the renderbuffer target is actually an
extension and not always available in GLES2.0. Instead we can use
glRenderbufferStorage and glFramebufferRenderbuffer for the same
result here as we do not later make use of the depth buffer as
a texture. This makes this more portable to more basic GLES2.0
implementations.

Signed-off-by: Andrew Davis <afd@ti.com>