cube-shadertoy: Do not assume 0 is default framebuffer number
The function draw_shadertoy() binds a texture framebuffer object for
rendering the shadertoy onto and then restores the old framebuffer.
But it assumes the old framebuffer object number was 0, which is usually
the case but that is not guaranteed. To be fully correct we should fetch
the current framebuffer number and restore to that.

This exposes another issue that init_shadertoy() rebinds the framebuffer
but never sets it back to the original value. This was masked by the
unconditional binding to 0 in draw_shadertoy(). Now that it doesn't always
rebind to 0, we never rebind to the original. So save and restore the
framebuffer number in init_shadertoy() also.

Signed-off-by: Andrew Davis <afd@ti.com>