cube-gears: Only set vertex attributes when needed
Without vertex array objects (not available in ES2.0), vertex attributes
are global and so should be setup, enabled, and then disabled around the
draw function that make use of them. This keeps state local to prevent
bugs.

The draw_gear() function already does this, but the cube drawing does not.
Move the vertex attribute setup out of init() and to draw(). This will
prevent any accidental overwriting of vertex attributes should any other
attribute location alias these in a different program (attribute locations
are per-program, but attribute indexes are global!).

Signed-off-by: Andrew Davis <afd@ti.com>