cube-video: Use new VBO for blit vertex data
The 3 cubes: gears, shadertoy, and video all use a second program for
rendering something other than a cube. For the first two they already
use a second VBO for the non-cube vertex data, but video cube aliases
the cube VBO to define vertices for the background rendering. While
this does work, it is easy to define exactly the vertex data we need
and put this in a separate VBO. This helps later if we want to better
split background rendering from cube face rendering.

Signed-off-by: Andrew Davis <afd@ti.com>