cube-video: Create new frame after setting up new decoder
When we get to the end of a video a NULL frame is returned, this signals
that we should rebuild the decoder to start a new video. The issue is
after making this new decoder we do not fetch a new frame, which means
a NULL frame EGLImage is set as the texture target. This is usually
harmless and results in a single frame of black video.

Instead fetch a new frame. While here also check that the new decoder
was created successfully.

Signed-off-by: Andrew Davis <afd@ti.com>